requirements.txt
Flask==3.0.3
fpdf==1.7.2
pillow==10.4.0
python-dotenv==1.0.1
python-resize-image==1.1.20
redis==5.1.1
requests==2.32.3
spotipy==2.24.0
urllib3==2.2.3